fix: refetch 함수 props로 넘기기

+++ client/src/components/Community/CommentListItem.jsx
@@ -2,7 +2,7 @@ import React, { useRef, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import { toast } from 'react-toastify';
 
-const CommentListItem = ({ comment, toggleReply, userId }) => {
+const CommentListItem = ({ comment, toggleReply, userId, refetch }) => {
   const [hidden, setHidden] = useState(true);
   const { category, id } = useParams();
   const modifyRef = useRef();
@@ -25,17 +25,26 @@ const CommentListItem = ({ comment, toggleReply, userId }) => {
         }),
       }
     );
-    window.location.reload();
+    setHidden(true);
+    refetch();
   };
 
   const clickDelete = (event) => {
-    fetch(
-      `http://localhost:8080/api/post/reply/${category}/${id}/${userId}/${comment.id}`,
-      {
-        method: 'DELETE',
-      }
-    );
-    window.location.reload();
+    if (confirm('정말 삭제할까용')) {
+      fetch(
+        `http://localhost:8080/api/post/reply/${category}/${id}/${userId}/${comment.id}`,
+        {
+          method: 'DELETE',
+        }
+      );
+      toast.success('삭제가 완료되었습니다.', {
+        autoClose: 500,
+        position: toast.POSITION.BOTTOM_CENTER,
+      });
+      setHidden(true);
+      // refetch();
+      window.location.reload();
+    }
   };
 
   return (
